+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 - The `composer` command now uses the `craftcms/cli` docker image and runs in the same network as the nitro sites. [#241](https://github.com/craftcms/nitro/issues/241)
 
+### Removed
+- The `--composer-version` flag on the `composer` command has been replaced with `--php-version`.
+
 ## 2.0.0-beta.5 - 2021-02-16
 
 ### Added